Add keyword quiet (logical) to vgd_get to suppress error message
Add test to above feature (get_quiet.ftn90)
Had to modif run_test.ksh to implement shell script testing. These tests must
be put in tests/ksh and must by dottable files. Control output for these tests
are in data_AIX/TXT data_Linux/TXT 

Add preffix (vgd) to messages

Add Reduce severity code

MSG_CRITICAL -> MSG_ERROR
MSG_INFO     -> MSG_INFO

Use s/r msg for messaging from librairy modelutils than may be obtained with 

. s.ssmuse.dot ...

and link with

 -libappl modelutils_base



git-svn-id: svn://mrbsvn/others/vcoord/trunk@854 198adb2e-9c79-420d-9e7c-cb98fabe6407

+++ tests/run_tests.ksh
@@ -7,6 +7,16 @@ else
   set -A tests $*
 fi
 
+#==============
+function Other_Tests {
+
+   result_o=' ok'
+   if [ ${test} = get_quiet ];then
+       . ksh/other_test.dot
+   fi
+
+}
+#==============
 # Compile all tests
 printf "Compiling tests\n"
 template=Makefile.tmpl
@@ -28,6 +38,7 @@ for test in ${tests[*]} ; do
 done
 printf " * All Builds Succeeded\n"
 
+#==============
 # Run all tests
 printf "Running tests\n"
 report_file=test_report.txt
@@ -37,13 +48,18 @@ for test in ${tests[*]} ; do
   printf "Testing ${test} ..."
   ${test} >${test}.out 2>&1
   result=`cat ${report_file}`
-  if [[ ${result} == ' ok' ]] ; then
+  Other_Tests
+  if [[ ${result} == ' ok' && ${result_o} == ' ok' ]] ; then
      passed=$((passed + 1))
      printf "${result}\n"
      rm -f ${test} ${test}.out
   else
      failed=$((failed + 1))
-     printf "${result} (see ${test}.out for details)\n"
+     if [ "${result}" != ' ok' ];then
+        printf "${result} (see ${test}.out for details)\n"
+     else
+        printf "   Shell script test failled, see reason above"
+     fi
   fi
   rm -f ${report_file}
 done
